cli/requirements.txt
snaptrade-python-sdk>=2.0.0
click>=8.1.0
python-dotenv>=1.0.0
